+++ developer/src/kmc-kmn/src/compiler/kmn-compiler-messages.ts
@@ -42,7 +42,7 @@ export const enum KmnCompilerMessageRanges {
 
 /*
   The messages in this class share the namespace with messages from
-  kmn_compiler_errors.h, which are generated by kmcmplib, so values below 0x1000
+  kmn_compiler_errors.h, which are generated by kmcmplib, so values below 0x900
   are reserved for kmcmplib messages.
 */
 export class CompilerMessages {
