+++ developer/src/kmc-kmn/src/main.ts
@@ -1,4 +1,4 @@
 
 export { KmnCompiler } from './compiler/compiler.js';
-export { KmnCompilerMessages, CompilerMessages } from './compiler/messages.js';
-export { KmwCompilerMessages } from './kmw-compiler/messages.js';
+export { KmnCompilerMessages, CompilerMessages } from './compiler/kmn-compiler-messages.js';
+export { KmwCompilerMessages } from './kmw-compiler/kmw-compiler-messages.js';
